Code reorganization for node registry REST API
- Compartmentalize actix-specific implementations behind the
`rest-api-actix` feature
- Separate resources into their own modules
