+++ libsplinter/src/node_registry/rest_api/resources/nodes.rs
@@ -1,0 +1,22 @@
+// Copyright 2018-2020 Cargill Incorporated
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::node_registry::Node;
+use crate::rest_api::paging::Paging;
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+pub struct ListNodesResponse {
+    pub data: Vec<Node>,
+    pub paging: Paging,
+}
